+++ pages/blog/index.tsx
@@ -1,5 +1,6 @@
 import { Layout } from "@components/layout";
 import { PostPanel } from "@components/panel";
+import { Parallax } from "@components/parallax";
 import { Section } from "@components/section";
 import { IPost } from "@util/interface";
 import { getBlogPosts } from "@util/notion";
@@ -13,13 +14,24 @@ const BlogPage: NextPage<BlockPageProps> = ({ posts }: BlockPageProps) => {
     return (
         <Layout title="Blog" description="My engineering blog">
             <Section>
-                <div className="flex flex-col 2xl:flex-row 2xl:columns-2 my-40 w-11/12 mx-auto">
-                    <div className="flex-shrink">
-                        <span className="font-clash text-5xl md:text-8xl font-light text-center mx-auto w-full">
-                            My Blog
-                        </span>
-                    </div>
-                    <div className="flex-grow px-2">
+                <div className="blog-section">
+                    <Parallax range={[0, -0.5]} className="blog-hero">
+                        <span className="blog-title">My Blog</span>
+                        <div className="introduction">
+                            <p>
+                                Lorem ipsum dolor sit amet, consectetur
+                                adipiscing elit. Ut sit amet quam in leo
+                                molestie elementum. Mauris eleifend pretium
+                                sapien. Aenean ac sagittis purus. Nam eget
+                                dictum magna. Integer molestie neque eu odio
+                                vehicula dapibus. Proin ullamcorper diam quis
+                                consectetur porttitor. Quisque id ex non mi
+                                efficitur dictum nec viverra ligula. Nam
+                                pulvinar quis diam non tempor. <br />
+                            </p>
+                        </div>
+                    </Parallax>
+                    <div className="blog-main">
                         <PostPanel posts={posts} />
                     </div>
                 </div>