+++ src/components/marquee/index.tsx
@@ -1,3 +1,0 @@
-export const Marquee: React.FC = () => {
-    return <div></div>;
-};