+++ src/components/panel/index.tsx
@@ -1,7 +1,8 @@
 import { PostCard, ProjectCard } from "@components/card";
 import { IPost, IProject } from "@util/interface";
-import { motion } from "framer-motion";
-import { PropsWithChildren } from "react";
+import { techStackDetails } from "@util/tech";
+import { AnimatePresence, motion } from "framer-motion";
+import { PropsWithChildren, useState } from "react";
 
 const PanelVariants = {
     hidden: {},
@@ -20,7 +21,7 @@ export const ProjectPanel: React.FC<ProjectPanelProps> = ({
             initial="hidden"
             whileInView="enter"
             variants={PanelVariants}
-            className="m-auto"
+            className="w-full md:w-3/4 ml-2"
         >
             {projects.map((project) => (
                 <ProjectCard key={project.id} {...project} />
@@ -49,5 +50,55 @@ export const PostPanel: React.FC<PostPanelProps> = ({
 };
 
 export const TechStack: React.FC = () => {
-    return <div>Python</div>;
+    const [isSelected, setSelected] = useState(techStackDetails[0]);
+
+    return (
+        <div className="tech-stack">
+            <div className="nav">
+                {techStackDetails.map((tech) => (
+                    <div
+                        key={tech.id}
+                        className={`
+                        nav-item 
+                        ${tech.id == isSelected.id ? "selected" : ""}
+                    `}
+                        onClick={() => setSelected(tech)}
+                    >
+                        {tech.acronym}
+                        {tech.id === isSelected.id ? (
+                            <motion.div
+                                className="highlight"
+                                layoutId="underline"
+                            />
+                        ) : null}
+                    </div>
+                ))}
+            </div>
+            <div>
+                <AnimatePresence exitBeforeEnter>
+                    <motion.div
+                        key={isSelected ? isSelected.domain : "empty"}
+                        animate={{ opacity: 1, y: 0 }}
+                        initial={{ opacity: 0, y: 20 }}
+                        exit={{ opacity: 0, y: -20 }}
+                        transition={{ duration: 0.15 }}
+                        className="main"
+                    >
+                        <span className="domain">{isSelected.domain}</span>
+                        <div className="tools">
+                            {isSelected.tools.map((Icon) => (
+                                <Icon
+                                    key={Icon.name}
+                                    className="hover:text-accent"
+                                />
+                            ))}
+                        </div>
+                        <div className="description">
+                            <p>{isSelected.description}</p>
+                        </div>
+                    </motion.div>
+                </AnimatePresence>
+            </div>
+        </div>
+    );
 };