+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.4] - 2022-01-30
+### Added
+- Stylesheets for all components necessary
+- Parallax Divs
+- TechStack Panel
+- Double Icon button
+### Changed
+- Section layout improved
+- Hero Section animations and layout
+- Colors
+- Footer content
+### Removed
+- Marquee component
+
 ## [0.3.3] - 2022-01-29
 ### Added
 - Stylesheet for blog components